Move profile-view-related stuff from UserBean to ProfileBean + Replace GeneralView-Variables by calling SecurityBean
