+++ src/main/webapp/pages/login.xhtml
@@ -10,7 +10,6 @@
     <ui:define name="title">#{msgs['login.title']}</ui:define>
     <ui:define name="header">#{msgs['login.header']}</ui:define>
     <ui:define name="content">
-
         <f:facet name="last">
             <h:outputStylesheet name="login.css" library="styles" />
         </f:facet>
@@ -34,7 +33,6 @@
                 </p:panelGrid>
             </h:form>
         </p:panel>
-
     </ui:define>
 </ui:composition>
