+++ src/main/webapp/pages/loginsuccess.xhtml
@@ -10,7 +10,6 @@
     <ui:define name="title">#{msgs['login.title']}</ui:define>
     <ui:define name="header">#{msgs['login.header']}</ui:define>
     <ui:define name="content">
-
         <f:facet name="last">
             <h:outputStylesheet name="login.css" library="styles" />
         </f:facet>
@@ -20,7 +19,6 @@
                 <h:outputText value="Willkommen zurück, #{securityBean.user.name}!" />
             </p:panelGrid>
         </p:panel>
-
     </ui:define>
 </ui:composition>
