+++ src/main/webapp/pages/my-profile.xhtml
@@ -11,33 +11,28 @@
     <ui:define name="title">#{msgs['user.title']}</ui:define>
     <ui:define name="header">#{msgs['user.header']}</ui:define>
     <ui:define name="content">
-        <f:facet name="last">
-            <h:outputStylesheet name="myProfile.css" library="styles" />
-        </f:facet>
+        <h:form id="form">
+            <f:facet name="last">
+                <h:outputStylesheet name="myProfile.css" library="styles" />
+            </f:facet>
 
-        <p:panel id="myProfilePanel" header="#{msgs['user.details']}" styleClass="panelGridCenterAndTop">
-
-            <h:form id="form">
-
-
-                #{generalView.getUserToDisplay()}
-
+            <p:panel id="myProfilePanel" header="#{msgs['user.details']}" styleClass="panelGridCenterAndTop">
                 <h:panelGrid style="width:100%" columns="2" cellpadding="80" columnClasses="description, detail" styleClass="panelGridCenter" >
                     <h:outputLabel value="#{msgs['label.name']}: " style="font-weight: bold"/>
-                    <h:outputText value="#{generalView.name}" />
+                    <h:outputText value="#{securityBean.user.name}" />
 
                     <h:outputLabel value="#{msgs['label.login_id']}: " style="font-weight: bold"/>
-                    <h:outputText value="#{generalView.userName}" />
+                    <h:outputText value="#{securityBean.user.loginID}" />
 
                     <h:outputLabel value="#{msgs['label.email']}: " style="font-weight: bold"/>
-                    <h:outputText value="#{generalView.email}" />
+                    <h:outputText value="#{securityBean.user.email}" />
 
                     <h:outputLabel value="#{msgs['label.company']}: " style="font-weight: bold"/>
-                    <h:outputText value="#{generalView.company}" />
+                    <h:outputText value="#{securityBean.user.company}" />
 
                     <h:outputLabel value="#{msgs['label.role']}: " style="font-weight: bold"/>
                     <p:outputPanel>
-                        <h:dataTable value="#{generalView.roles}" var="role"><f:facet name="Role">Status</f:facet>
+                        <h:dataTable value="#{securityBean.user.roles}" var="role"><f:facet name="Role">Status</f:facet>
                             <h:column>
                                 #{role.name}
                             </h:column>
@@ -47,7 +42,7 @@
                 <h:panelGrid style="width:100%" columns="1" cellpadding="80" columnClasses="description, detail" styleClass="commandButton" >
                     <p:growl id="message" showDetail="true" />
 
-                    <p:commandButton value="Profil löschen" action="#{userBean.deleteAccount()}" update="message">
+                    <p:commandButton value="Profil löschen" action="#{profileBean.deleteAccount()}" update="message">
                         <p:confirm header="Achtung" message="Möchten Sie Ihr Profil wirklich löschen?"/>
                     </p:commandButton>
 
@@ -56,8 +51,8 @@
                         <p:commandButton value="Nein" type="button" styleClass="ui-confirmdialog-no"/>
                     </p:confirmDialog>
                 </h:panelGrid>
-            </h:form>
-        </p:panel>
+            </p:panel>
+        </h:form>
     </ui:define>
 </ui:composition>
 </html>
